% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop-transformation.R
\name{build_transformed_function}
\alias{build_transformed_function}
\title{Construct the expression for a transformed function body.}
\usage{
build_transformed_function(fun_expr, fun_name, fun)
}
\arguments{
\item{fun_expr}{The original function body.}

\item{fun_name}{The name of the recursive function.}

\item{fun}{The actual function definition.}
}
\value{
The body of the transformed function.
}
\description{
This is where the loop-transformation is done. This function translates
the body of a recursive function into a looping function.
}
