% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop-transformation.R
\name{can_call_be_transformed}
\alias{can_call_be_transformed}
\title{Tests if a call object can be transformed.}
\usage{
can_call_be_transformed(call_name, call_arguments, fun_name, fun_call_allowed,
  cc)
}
\arguments{
\item{call_name}{Name (function) of the call.}

\item{call_arguments}{The call's arguments}

\item{fun_name}{The name of the recursive function we want to transform}

\item{fun_call_allowed}{Whether a recursive call is allowed at this point}

\item{cc}{Current continuation to abort if a transformation is not possible}
}
\value{
TRUE, if the expression can be transformed. Invokes \code{cc} otherwise.
}
\description{
Tests if a call object can be transformed.
}
