% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop-transformation.R
\name{can_transform_rec}
\alias{can_transform_rec}
\title{Recursive call for testing if an expression can be transformed into a looping
tail-recursion.}
\usage{
can_transform_rec(expr, fun_name, fun_call_allowed, cc)
}
\arguments{
\item{expr}{The expression to test}

\item{fun_name}{The name of the recursive function we want to transform}

\item{fun_call_allowed}{Whether a recursive call is allowed at this point}

\item{cc}{Current continuation, used to escape if the expression cannot be transformed.}
}
\value{
TRUE, if the expression can be transformed. Invokes \code{cc} otherwise.
}
\description{
Recursive call for testing if an expression can be transformed into a looping
tail-recursion.
}
