% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop-transformation.R
\name{capture_closure}
\alias{capture_closure}
\title{Captures the closure in a function environment}
\usage{
capture_closure(.func)
}
\arguments{
\item{.func}{The function to capture a closure for.}
}
\description{
This is necessary to implement continuation-passing-style
tail-recursion.
}
