% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop-transformation.R
\name{handle_recursive_returns}
\alias{handle_recursive_returns}
\title{Handle the actual recursive calls}
\usage{
handle_recursive_returns(expr, info)
}
\arguments{
\item{expr}{An expression to transform}

\item{info}{Information passed along the transformations.}
}
\value{
A modified expression.
}
\description{
Handle the actual recursive calls
}
