% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop-transformation.R
\name{handle_recursive_returns_call}
\alias{handle_recursive_returns_call}
\title{Handles the actual recursive returns}
\usage{
handle_recursive_returns_call(call_expr, info)
}
\arguments{
\item{call_expr}{The call to modify.}

\item{info}{Information passed along with transformations.}
}
\value{
A modified expression.
}
\description{
This function dispatches on a call object to set the context of recursive
expression modifications.
}
