% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop-transformation.R
\name{loop_transform}
\alias{loop_transform}
\title{Transform a function from recursive to looping.}
\usage{
loop_transform(fun, byte_compile = TRUE)
}
\arguments{
\item{fun}{The function to transform. Must be provided as a bare name.}

\item{byte_compile}{Flag specifying whether to compile the function after
transformation.}
}
\description{
Since this function needs to handle recursive functions, it needs to know the
name of its input function, so this must be provided as a bare symbol.
}
