% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop-transformation.R
\name{make_returns_explicit}
\alias{make_returns_explicit}
\title{Make exit points into explicit calls to return.}
\usage{
make_returns_explicit(expr, in_function_parameter, info)
}
\arguments{
\item{expr}{An expression to transform}

\item{in_function_parameter}{Is the expression part of a parameter to a function call?}

\item{info}{Information passed along the transformations.}
}
\value{
A modified expression.
}
\description{
Make exit points into explicit calls to return.
}
