% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop-transformation.R
\name{make_returns_explicit_call}
\alias{make_returns_explicit_call}
\title{Make exit points into explicit calls to return.}
\usage{
make_returns_explicit_call(call_expr, in_function_parameter, info)
}
\arguments{
\item{call_expr}{The call to modify.}

\item{in_function_parameter}{Is the expression part of a parameter to a function call?}

\item{info}{Information passed along with transformations.}
}
\value{
A modified expression.
}
\description{
This function dispatches on a call object to set the context of recursive
expression modifications.
}
