% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop-transformation.R
\name{returns_to_escapes}
\alias{returns_to_escapes}
\title{Make calls to return into calls to escapes.}
\usage{
returns_to_escapes(expr, info)
}
\arguments{
\item{expr}{An expression to transform}

\item{info}{Information passed along the transformations.}
}
\value{
A modified expression.
}
\description{
Make calls to return into calls to escapes.
}
