% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop-transformation.R
\name{returns_to_escapes_call}
\alias{returns_to_escapes_call}
\title{Make calls to return into calls to escapes.}
\usage{
returns_to_escapes_call(call_expr, info)
}
\arguments{
\item{call_expr}{The call to modify.}

\item{info}{Information passed along with transformations.}
}
\value{
A modified expression.
}
\description{
This function dispatches on a call object to set the context of recursive
expression modifications.
}
