% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop-transformation.R
\name{simplify_nested_blocks}
\alias{simplify_nested_blocks}
\title{Simplify nested code-blocks.}
\usage{
simplify_nested_blocks(expr)
}
\arguments{
\item{expr}{The expression to rewrite}
}
\value{
The new expression
}
\description{
If a call is \code{\{} and has a single expression inside it, replace it with that expression.
}
