% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop-transformation.R
\name{simplify_returns}
\alias{simplify_returns}
\title{Remove return(return(...)) expressions}
\usage{
simplify_returns(expr, info)
}
\arguments{
\item{expr}{An expression to transform}

\item{info}{Information passed along the transformations.}
}
\value{
A modified expression.
}
\description{
Remove return(return(...)) expressions
}
