% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop-transformation.R
\name{transform_recursive_calls}
\alias{transform_recursive_calls}
\title{Translate all return(<recursive-function-call>) expressions into
a block that assigns the parameters to local variables.}
\usage{
transform_recursive_calls(expr, fun_name, fun)
}
\arguments{
\item{expr}{The expression to rewrite}

\item{fun_name}{The name of the recursive function we are rewriting}

\item{fun}{The actual function -- we use this for the call to `match.call`.}
}
\value{
The rewritten expression
}
\description{
Translate all return(<recursive-function-call>) expressions into
a block that assigns the parameters to local variables.
}
