% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop-transformation.R
\name{translate_recursive_call_into_next}
\alias{translate_recursive_call_into_next}
\title{Translate a return(<recursive-function-call>) expressions into
a block that assigns the parameters to local variables and call `continue`.}
\usage{
translate_recursive_call_into_next(recursive_call, fun)
}
\arguments{
\item{recursive_call}{The call object where we get the parameters}

\item{fun}{The actual function -- we use this for the call to `match.call`.}
}
\value{
The rewritten expression
}
\description{
Translate a return(<recursive-function-call>) expressions into
a block that assigns the parameters to local variables and call `continue`.
}
