% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-defined-transformations.R
\name{user_transform}
\alias{user_transform}
\title{Apply user transformations depths-first.}
\usage{
user_transform(fn)
}
\arguments{
\item{fn}{The actual function to transform.}
}
\value{
Rewritten expression
}
\description{
Apply user transformations depths-first.
}
