% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop-transformation.R
\name{can_transform}
\alias{can_transform}
\title{Tests if a function, provided by its name, can be transformed.}
\usage{
can_transform(fun)
}
\arguments{
\item{fun}{The function to check. Must be provided by its (bare symbol) name.}
}
\description{
Tests if a function, provided by its name, can be transformed.
}
\examples{
factorial <- function(n)
    if (n <= 1) 1 else n * factorial(n - 1)
factorial_acc <- function(n, acc = 1)
    if (n <= 1) acc else factorial_acc(n - 1, n * acc)
can_transform(factorial) # FALSE
can_transform(factorial_acc) # TRUE

}
