% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-defined-transformations.R
\name{user_transform}
\alias{user_transform}
\title{Apply user transformations depths-first.}
\usage{
user_transform(expr, fun = expr, env = rlang::caller_env())
}
\arguments{
\item{expr}{The expression to transform -- typically a function body.}

\item{fun}{The actual function to transform.}

\item{env}{The environment where functions can be found.}
}
\value{
Rewritten expression
}
\description{
Apply user transformations depths-first.
}
