% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-defined-transformations.R
\name{user_transform_rec}
\alias{user_transform_rec}
\title{Apply user transformations depths-first.}
\usage{
user_transform_rec(expr, env = rlang::caller_env())
}
\arguments{
\item{expr}{The expression to transform -- typically a function body.}

\item{env}{The environment where functions can be found.}
}
\value{
Rewritten expression
}
\description{
The difference between this function and \code{\link{user_transform}} is that the this
function does not perform type checks before calling recursively while \code{\link{user_transform}}
does.
}
